% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrolytes.R
\name{adjust_sodium_for_glucose}
\alias{adjust_sodium_for_glucose}
\title{Adjust sodium for hyperglycaemia}
\usage{
adjust_sodium_for_glucose(sodium, glucose, method = "Hillier")
}
\arguments{
\item{sodium}{Measured sodium level (mmol/l).}

\item{glucose}{Measured glucose level (mmol/l).}

\item{method}{Formula. Options are: "Hillier" (default), "Katz".}
}
\value{
Glucose-adjusted sodium level (mmol/l).
}
\description{
Hyperglycaemia can lead to hypertonic hyponatriemia. Using observational
data, the following formula predicts sodium levels after correction of
glucose: \eqn{\text{Na}_{predicted} = \text{Na}_\text{measured} + \kappa
\times \frac{\text{glucose} - 5.6}{5.6}}, where \eqn{\kappa} is 1.6 when
derived from the publication by Katz and 2.4 when derived from more recent
work by Hillier (default).
}
\section{Caveats}{
 It's just a prediction, and the choice between both
formulae may result in quite different predictions.
}

\references{
\href{https://doi.org/10.1056/NEJM197310182891607}{Katz, M. A. Hyperglycemia-Induced Hyponatremia — Calculation of Expected Serum Sodium Depression. N Engl J Med 289, 843–844 (1973).} and \href{https://doi.org/10.1016/s0002-9343(99)00055-8}{Hillier, T. A., Abbott, R. D. & Barrett, E. J. Hyponatremia: evaluating the correction factor for hyperglycemia. The American Journal of Medicine 106, 399–403 (1999).}.
}
\seealso{
\code{\link[units:set_units]{units::set_units()}}, \code{\link[units:drop_units]{units::drop_units()}}
}
