% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{calculate_bsa}
\alias{calculate_bsa}
\alias{calculate_bsa_monsteller}
\alias{calculate_bsa_du_bois}
\title{Calculate the body surface area (BSA).}
\usage{
calculate_bsa(weight, height, method = c("Monsteller", "Du Bois"))

calculate_bsa_monsteller(weight, height)

calculate_bsa_du_bois(weight, height)
}
\arguments{
\item{weight}{Weight in kg.}

\item{height}{Height in cm.}

\item{method}{Formula to calculate BSA, either Monsteller (default) or Du Bois.
(default).}
}
\value{
BSA in m^2.
}
\description{
Calculate the body surface area (BSA), according to Monsteller (see
\url{https://doi.org/10.1056\%2FNEJM198710223171717}) or Du Bois (see
\url{https://doi.org/10.1001\%2Farchinte.1916.00080130010002}).
}
\section{Functions}{
\itemize{
\item \code{calculate_bsa_monsteller}: Calculate the body surface area (BSA) according to Monsteller.

\item \code{calculate_bsa_du_bois}: Calculate the body surface area (BSA) according to Du Bois
}}

\keyword{internal}
