% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_pressure.R
\name{calculate_map}
\alias{calculate_map}
\title{Calculate the mean arterial pressure (MAP).}
\usage{
calculate_map(systolic_blood_pressure, diastolic_blood_pressure)
}
\arguments{
\item{systolic_blood_pressure}{Systolic blood pressure (mmHg).}

\item{diastolic_blood_pressure}{Diastolic blood pressure (mmHg).}
}
\value{
MAP (mmHg).
}
\description{
The mean arterial pressure is determined for 1/3 by systolic and 2/3 by diastolic blood pressure.
}
\section{Caveats}{
 None at this time.
}

\references{
None.
}
\seealso{
\code{\link[units]{set_units}}, \code{\link[units]{drop_units}}
}
