% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtc.R
\name{calculate_qtc_bazett}
\alias{calculate_qtc_bazett}
\alias{calculate_qtc_fridericia}
\title{Calculate the corrected QT intevral (QTc).}
\usage{
calculate_qtc_bazett(qt, heart_rate)

calculate_qtc_fridericia(qt, heart_rate)
}
\arguments{
\item{qt}{QT interval in ms.}

\item{heart_rate}{Heart rate in beats/min.}
}
\value{
QTc in ms
}
\description{
Calculate the corrected QT interval (QTc) of a electrocardiogram, according to Bazett (Bazett, 1920) or Fridericia (Fridericia, 1920).
}
