% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrolytes.R
\name{correct_calcium_for_albumin}
\alias{correct_calcium_for_albumin}
\title{Correct calcium for albumin level.}
\usage{
correct_calcium_for_albumin(calcium, albumin, normal_albumin = 40)
}
\arguments{
\item{calcium}{Measured (uncorrected) total calcium level (mmol/l).}

\item{albumin}{Measured albumin level (g/l).}

\item{normal_albumin}{Normal value of albumin (g/l) (40 by default).}
}
\value{
Albumin-corrected calcium level (mmol/l).
}
\description{
Most of the protein-bound calcium is bound to albumin. Lower albumin will lower total calcium levels, but the ionised calcium level will not be lowered that much. Therefore, total calcium levels should be corrected for albumin, since in a state of
hypoalbuminemia the (ionised, relevant) calcium level is underestimated. An alternative approach
is measuring ionised calcium levels directly.
}
\details{
The formula used here is (calcium + 0.025 * (normal albumin - measured albumin)).
}
\section{Caveats}{
 Corrections for albumin are quite poor substitutes for measuring ionised calcium.
}

\references{
Calcium (Bushinsky; 1998): \url{https://doi.org/10.1016/S0140-6736(97)12331-5}.
}
\seealso{
\code{\link[units]{set_units}}, \code{\link[units]{drop_units}}
}
