% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{estimate_bsa}
\alias{estimate_bsa}
\title{Estimate the body surface area (BSA)}
\usage{
estimate_bsa(weight, height, method = "Monsteller")
}
\arguments{
\item{weight}{Weight (kg).}

\item{height}{Height (cm).}

\item{method}{Formula to estimate BSA. Options are: "Monsteller" (default),
"Du Bois".}
}
\value{
BSA (\eqn{m^2}).
}
\description{
Estimate the body surface area (BSA), according to Monsteller or Du Bois.
}
\section{Caveats}{
 None at this time.
}

\references{
\href{https://doi.org/10.1056\NEJM198710223171717}{Simplified Calculation of Body-Surface Area. N Engl J Med 317, 1098–1098 (1987).} and \href{https://doi.org/10.1001/archinte.1916.00080130010002}{Du BOIS, D. CLINICAL CALORIMETRY. Arch Intern Med (Chic) XVII, 863 (1916).}.
}
\seealso{
\code{\link[units:set_units]{units::set_units()}}, \code{\link[units:drop_units]{units::drop_units()}}
}
