% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renal_function.R
\name{estimate_gfr_cockcroft}
\alias{estimate_gfr_cockcroft}
\title{Estimate creatinine clearance and glomerular filtration rate using the Cockcroft-Gault formula}
\usage{
estimate_gfr_cockcroft(creatinine, age, is_female, weight)
}
\arguments{
\item{creatinine}{Creatinine (\eqn{\text{μmol} \text{l}^{-1}}).}

\item{age}{Age (years).}

\item{is_female}{\code{TRUE} if patient is female, \code{FALSE} if patient is
male.}

\item{weight}{Weight (kg).}
}
\value{
eGFR (\eqn{\text{ml} \text{min}^{-1}}), or \code{NA} if any parameters are \code{NA}.
}
\description{
The formula is \eqn{\frac{(140 - \text{age}) \times \text{weight}}{0.81 \times \text{creatinine}} \times S}, where \eqn{S} is 1 for males and 0.85 for females.
}
\section{Caveats}{

Creatinine clearance may overestimate glomerular filtration rate in kidney injury. In non-ideal body composition, Cockcroft-Gault formula may be (increasingly) inaccurate. Note that the eGFR provided by this formula is \emph{not} normalised for a standardised body surface area of 1.73 \eqn{m^2} - to achieve this, divide the eGFR by the patients body surface area (using your method of choice) and multiply by 1.73 \eqn{m^2}.
}

\references{
\href{https://doi.org/10.1159/000180580}{Cockcroft, D. W. & Gault, H. Prediction of Creatinine Clearance from Serum Creatinine. Nephron 16, 31–41 (1976).}
}
\seealso{
\code{\link[units:set_units]{units::set_units()}}, \code{\link[units:drop_units]{units::drop_units()}}
}
