% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renal_function.R
\name{estimate_gfr_schwartz}
\alias{estimate_gfr_schwartz}
\title{Estimate creatinine clearance and glomerular filtration rate using the Schwartz formula}
\usage{
estimate_gfr_schwartz(creatinine, height)
}
\arguments{
\item{creatinine}{Creatinine (\eqn{\text{μmol} \text{l}^{-1}}).}

\item{height}{Height (cm).}
}
\value{
eGFR (\eqn{\text{ml} \text{min}^{-1} 1.73 \text{m}^{-2}}), or \code{NA} if any parameters are \code{NA}.
}
\description{
The Schwartz formula is used in children: \eqn{\frac{36.2 \times \text{height}}{\text{creatinine}}}.
}
\section{Caveats}{

This is the revised (2009) formula with the same factor \eqn{\kappa} for all age categories: 36.2. The returned value is based on a standardised body surface area of 1.73 \eqn{m^2}.
}

\references{
\href{https://doi.org/10.2215/CJN.01640309}{Schwartz, G. J. & Work, D. F. Measurement and Estimation of GFR in Children and Adolescents. CJASN 4, 1832–1843 (2009).}
}
\seealso{
\code{\link[units:set_units]{units::set_units()}}, \code{\link[units:drop_units]{units::drop_units()}}
}
