% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{estimate_ibw}
\alias{estimate_ibw}
\title{Estimate ideal body weight (IBW)}
\usage{
estimate_ibw(height, is_female)
}
\arguments{
\item{height}{Height (cm).}

\item{is_female}{\code{TRUE} if patient is female, \code{FALSE} if patient is
male.}
}
\value{
IBW (kg), or \code{NA} if any parameters are \code{NA}.
}
\description{
Ideal body weight is estimated using the Devine formulae:
for males \eqn{\text{IBW} = 50 + 0.9 (\text{height} - 152)} and
for females \eqn{\text{IBW} = 45.5 + 0.9 (\text{height} - 152)}
with IBW in kg en height in m.
}
\section{Caveats}{
 This provides estimates for adults only.
}

\references{
\href{https://doi.org/10.1177/106002807400801104}{McCarron, M. M. & Devine, B. J. Clinical Pharmacy: Case Studies. Drug Intelligence & Clinical Pharmacy 8, 650–655 (1974).}
}
\seealso{
\code{\link[units:set_units]{units::set_units()}}, \code{\link[units:drop_units]{units::drop_units()}}
}
