% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_pressure.R
\name{estimate_map}
\alias{estimate_map}
\title{Estimate the mean arterial pressure (MAP)}
\usage{
estimate_map(sbp, dbp)
}
\arguments{
\item{sbp}{Systolic blood pressure (mmHg).}

\item{dbp}{Diastolic blood pressure (mmHg).}
}
\value{
MAP (mmHg).
}
\description{
The mean arterial pressure is estimated using the formula:
\eqn{\text{MAP} \approxeq P_{\text{diastolic}} + \frac{1}{3} (P_{\text{systolic}} - P_{\text{diastolic}})}.
}
\section{Caveats}{
 None at this time.
}

\references{
None.
}
\seealso{
\code{\link[units]{set_units}}, \code{\link[units]{drop_units}}
}
