% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_pressure.R
\name{estimate_mean_arterial_pressure}
\alias{estimate_mean_arterial_pressure}
\title{Estimate the mean arterial pressure (MAP)}
\usage{
estimate_mean_arterial_pressure(
  systolic_blood_pressure,
  diastolic_blood_pressure
)
}
\arguments{
\item{systolic_blood_pressure}{Systolic blood pressure (mmHg).}

\item{diastolic_blood_pressure}{Diastolic blood pressure (mmHg).}
}
\value{
MAP (mmHg).
}
\description{
The mean arterial pressure is estimated using the formula:
\eqn{\text{MAP} \approxeq P_{\text{diastolic}} + \frac{1}{3} (P_{\text{systolic}} - P_{\text{diastolic}})}.
}
\section{Caveats}{
 None at this time.
}

\references{
None.
}
\seealso{
\code{\link[units]{set_units}}, \code{\link[units]{drop_units}}
}
