% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pneumonia.R
\name{score_curb65}
\alias{score_curb65}
\title{Score community-acquired pneumonia severity using CURB-65}
\usage{
score_curb65(confusion, urea, respiratory_rate, sbp, dbp, age)
}
\arguments{
\item{confusion}{Mental confusion (\code{TRUE} or \code{FALSE}).}

\item{urea}{Blood urea nitrogen (mmol/l).}

\item{respiratory_rate}{Respiratory rate (/min).}

\item{sbp}{Systolic blood pressure (mmHg).}

\item{dbp}{Diastolic blood pressure (mmHg).}

\item{age}{Age (years).}
}
\value{
CURB-65 score (points), or \code{NA} if any parameters are \code{NA}.
}
\description{
The CURB-65 community-acquired pneumonia severity score is used to estimate
mortality.
}
\section{Caveats}{
 None at this time.
}

\references{
\href{https://doi.org/10.1136/thorax.58.5.377}{Lim, W. S. Defining community acquired pneumonia severity on presentation to hospital: an international derivation and validation study. Thorax 58, 377–382 (2003).}.
}
