% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepsis.R
\name{score_qsofa}
\alias{score_qsofa}
\title{Score qSOFA in sepsis}
\usage{
score_qsofa(gcs, respiratory_rate, sbp)
}
\arguments{
\item{gcs}{Glasgow Coma Scale.}

\item{respiratory_rate}{Respiratory rate (/min).}

\item{sbp}{Systolic blood pressure (mmHg).}
}
\value{
qSOFA score (points), or \code{NA} if any parameters are \code{NA}.
}
\description{
The quick SOFA (Sequential Organ Failure Assessment) score is used to estimate mortality in sepsis.
}
\section{Caveats}{
 Note that the original publications used GCS < 14 as
'altered mental state' but the current
\href{https://qsofa.org/what.php}{qSOFA website} promotes GCS < 15 as
cutoff. Here, GCS   < 15 is used as 'altered mental state'. Note that the
actual GCS does not matter for the qSOFA: if you only know if there was a
altered metal state, use \code{gcs = 15} for normal metal state and for
example \code{gcs = 1} for any altered metal state.
}

\references{
\href{https://doi.org/10.1001/jama.2016.0288}{Seymour, C. W. et al. Assessment of Clinical Criteria for Sepsis. JAMA 315, 762 (2016).}.
}
