% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Calculate the body mass index (BMI, Quetelet index).}
\usage{
calculate_bmi(weight, length)
}
\arguments{
\item{weight}{Weight in kg.}

\item{length}{Length in cm.}
}
\value{
BMI in kg/m^2.
}
\description{
Calculate the body mass index (BMI, Quetelet index).
}
