% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Calculate the body mass index (BMI, Quetelet index).}
\usage{
calculate_bmi(weight, height)
}
\arguments{
\item{weight}{Weight (kg).}

\item{height}{Height (cm).}
}
\value{
BMI (\eqn{kg m^-2}).
}
\description{
Calculate the body mass index (BMI, Quetelet index).
}
\seealso{
\code{\link[units]{set_units}}, \code{\link[units]{drop_units}}
}
