% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Calculate the body mass index (BMI, Quetelet index)}
\usage{
calculate_bmi(weight, height)
}
\arguments{
\item{weight}{Weight (kg).}

\item{height}{Height (cm).}
}
\value{
BMI (\eqn{\text{kg} m^{-2}}), or \code{NA} if any parameters are \code{NA}.
}
\description{
Body mass index is calculated using the formula: \eqn{\text{BMI} =
\text{weight} / \text{height}^2} with weight in kg en height in m.
}
\section{Caveats}{
 None at this time.
}

\references{
None.
}
\seealso{
\code{\link[units:set_units]{units::set_units()}}, \code{\link[units:drop_units]{units::drop_units()}}
}
