% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{calculate_bsa}
\alias{calculate_bsa}
\title{Calculate Body surface area (BSA).}
\usage{
calculate_bsa(weight, length)
}
\arguments{
\item{weight}{Weight in kg.}

\item{length}{Length in cm.}
}
\value{
BSA in m^2.
}
\description{
Calculate the body surface area (BSA), according to Monsteller (Monsteller, 1987).
}
