% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_pressure.R
\name{calculate_map}
\alias{calculate_map}
\title{Calculate the mean arterial pressure (MAP).}
\usage{
calculate_map(sbp, dbp)
}
\arguments{
\item{sbp}{Systolic blood pressure (mmHg).}

\item{dbp}{Diastolic blood pressure (mmHg).}
}
\value{
MAP (mmHg).
}
\description{
Mean arterial pressure = 1/3(SBP) + 2/3(DBP).
}
\section{Caveats}{
 None at this time.
}

\references{
None.
}
\seealso{
\code{\link[units]{set_units}}, \code{\link[units]{drop_units}}
}
