% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtc.R
\name{calculate_qtc}
\alias{calculate_qtc}
\alias{calculate_qtc_bazett}
\alias{calculate_qtc_fridericia}
\title{Calculate the corrected QT interval (QTc).}
\usage{
calculate_qtc(qt, heart_rate, method = "Fridericia")

calculate_qtc_bazett(qt, heart_rate)

calculate_qtc_fridericia(qt, heart_rate)
}
\arguments{
\item{qt}{QT interval (ms).}

\item{heart_rate}{Heart rate (\eqn{min^-1}).}

\item{method}{Formula for QTc. Options are: "Fridericia" (default),
"Bazzett".}
}
\value{
QTc (ms).
}
\description{
Calculate the corrected QT interval (QTc) of a electrocardiogram, according
to Bazett (Bazett, 1920) or Fridericia (Fridericia, 1920).
}
\section{Functions}{
\itemize{
\item \code{calculate_qtc_bazett}: QTc according to Bazett.

\item \code{calculate_qtc_fridericia}: QTc according to Fridericia.
}}

\seealso{
\code{\link[units]{set_units}}, \code{\link[units]{drop_units}}
}
