% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtc.R
\name{calculate_qtc}
\alias{calculate_qtc}
\title{Calculate the corrected QT interval (QTc)}
\usage{
calculate_qtc(qt, heart_rate, method = "Fridericia")
}
\arguments{
\item{qt}{QT interval (ms).}

\item{heart_rate}{Heart rate (/min).}

\item{method}{Formula for QTc. Options are: "Fridericia" (default),
"Bazzett".}
}
\value{
QTc (ms), or \code{NA} if any parameters are \code{NA}.
}
\description{
Calculate the corrected QT interval (QTc) of a electrocardiogram, according
to Bazett (Bazett, 1920) or Fridericia (Fridericia, 1920).
}
\seealso{
\code{\link[units:set_units]{units::set_units()}}, \code{\link[units:drop_units]{units::drop_units()}}
}
