% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{estimate_bsa}
\alias{estimate_bsa}
\title{Estimate the body surface area (BSA)}
\usage{
estimate_bsa(weight, height, method = "Monsteller")
}
\arguments{
\item{weight}{Weight (kg).}

\item{height}{Height (cm).}

\item{method}{Formula to estimate BSA. Options are: "Monsteller" (default),
"Du Bois".}
}
\value{
BSA (\eqn{m^2}).
}
\description{
Estimate the body surface area (BSA), according to Monsteller or Du Bois.
}
\section{Caveats}{
 None at this time.
}

\references{
Monsteller: \url{https://doi.org/10.1056\%2FNEJM198710223171717};
Du Bois: \url{https://doi.org/10.1001\%2Farchinte.1916.00080130010002}.
}
\seealso{
\code{\link[units]{set_units}}, \code{\link[units]{drop_units}}
}
