% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic_syndrome.R
\name{has_metabolic_syndrome_atpiii}
\alias{has_metabolic_syndrome_atpiii}
\title{Determine presence of metabolic syndrome using updated NCEP ATPIII criteria.}
\usage{
has_metabolic_syndrome_atpiii(
  is_female,
  waist_circumference = NA,
  systolic_blood_pressure = NA,
  diastolic_blood_pressure = NA,
  triglycerides = NA,
  hdl_cholesterol = NA,
  glucose = NA,
  is_fasting_blood_sample = FALSE,
  has_antihypertensive_drug = FALSE,
  has_lipid_drug = FALSE,
  has_glucose_drug = FALSE
)
}
\arguments{
\item{is_female}{\code{TRUE} if patient is female, \code{FALSE} if patient is a man.}

\item{waist_circumference}{Waist circumference (cm) (\code{NA} by default).}

\item{systolic_blood_pressure}{Systolic blood pressure (mmHg) (\code{NA} by default).}

\item{diastolic_blood_pressure}{Systolic blood pressure (mmHg) (\code{NA} by default).}

\item{triglycerides}{Triglycerides in (mmol/l) (\code{NA} by default).}

\item{hdl_cholesterol}{HDL cholesterol (mmol/l) (\code{NA} by default).}

\item{glucose}{Glucose (mmol/l) (\code{NA} by default).}

\item{is_fasting_blood_sample}{\code{TRUE} if fasting blood sample, \code{FALSE} (default) if not or
unknown.}

\item{has_antihypertensive_drug}{\code{TRUE} if patient is on antihypertensive drug
treatment, otherwise \code{FALSE} (default).}

\item{has_lipid_drug}{\code{TRUE} if patient is on fibrate, nicotinic acid or
statin, otherwise \code{FALSE} (default).}

\item{has_glucose_drug}{\code{TRUE} if patient is on drug treatment for elevated
glucose, otherwise \code{FALSE} (default).}
}
\value{
\code{TRUE} if patient has the metabolic syndrome, \code{FALSE} is not, \code{NA} if
  indetermined.
}
\description{
Determines whether an individual has the metabolic syndrome based on the
definition of the NCEP ATPIII criteria, updated in 2005 by Grundy et al.
}
\section{Caveats}{

We do include use of a statin as a drug for HDL cholesterol and
triglicerides, even though the original statement is unclear about this point
(only including niacin and fibrates).

According to the definition, only glucose needs to be from a fasting blood
sample (and not triglycerides or HDL cholesterol). The glucose trait is
determined in a fail-fast way: if you do not explicitly set \code{is_fasting_blood_sample = TRUE}, the glucose measurement is considered to be
non-fastening and is ignored.

This function explicitly handles situations of missing data. For example,
with blood pressure 120/80 mmHg and antihypertensive medication, a patient is
still considered hypertensive. With blood pressure \code{NA}/90 mmHg without
antihypertensive medication, a patient is still considered hypertensive. A
patient with blood pressure \code{NA}/\code{NA} mmHg and no antihypertensive medication is
considered \code{NA} for the trait hypertension, but if a patient has \code{NA}/\code{NA} mmHg but
does use antihypertensive medication, the patient is considered to have
hypertension (even though measurements are missing).
}

\references{
Updated NCEP ATPIII criteria (Grundy et al., 2005):
\url{https://doi.org/10.1161/circulationaha.105.169404}.
}
